+++ SimulationEngineLib/cpp/universe/Universe.h
@@ -11,11 +11,17 @@ enum Usage { UNDEFINED, TRUE, FALSE };
 
 class Universe {
 public:
+	// deltaTime is how much simulated time one step advances, in seconds.
+	// A float rather than a whole number of seconds: the step is the run
+	// length divided by the frames asked for, which is a fraction of a
+	// second as soon as the run is short or the frame count high, and
+	// truncating that to an integer either changes the run silently or
+	// collapses it to a zero-length step.
 	Universe(
 		std::vector<Particle*> particles, 
 		std::vector<std::shared_ptr<Law>> laws, 
 		const std::shared_ptr<SimulationOutput> output, 
-		unsigned int deltaTime,
+		float deltaTime,
 		unsigned long endTime,
 		Usage useGpu = UNDEFINED
 	);
@@ -26,7 +32,7 @@ public:
 protected:
 	std::vector<std::shared_ptr<Law>> laws;
 	const std::shared_ptr<SimulationOutput> output;
-	unsigned int deltaTime;
+	float deltaTime;
 	unsigned long endTime;
 	Usage useGpu;
