+++ SimulationEngineLib/shared/law/gravity/helper/NewtonGravityHelper.cu
@@ -17,7 +17,7 @@ Vector3D<float> getAcceleration(float mass, Vector3D<float> radiusComponent) {
 #endif
 void runOnParticle(
 	Particle* p1, Vector3D<float> acceleration, 
-	unsigned int deltaTime,
+	float deltaTime,
 	MomentumService* momentumService
 ) {
 	p1->velocity = momentumService->getVelocityPlusAcceleration(
