+++ SimulationEngineLib/shared/service/momentum/newton/NewtonMomentumService.cu
@@ -6,7 +6,7 @@ __device__ __host__
 Vector3D<float> NewtonMomentumService::getVelocityPlusAcceleration(
     float mass,
     Vector3D<float> acceleration, 
-    unsigned int deltaTime,
+    float deltaTime,
     Vector3D<float> velocity
 ) {
     return velocity + acceleration * deltaTime;
